+++ spike/version_rev.py
@@ -10,8 +10,8 @@ except:
     revision = "--not determined--"
 ProgramName = 'SPIKE'
 VersionName = 'Development version'
-version = '0.99.25'
-rev_date = '11-06-2021'
+version = '0.99.26'
+rev_date = '05-07-2021'
 
 def report():
     "prints version name when SPIKE starts"